REfactor e iliminando archivos inecesarios en los perfiles
